accounting: move config into its own file
